Print pointers as pointers, align specifiers wherever we notice they're incorrect.

+++ solving/move_effect_journal.c
@@ -136,7 +136,7 @@ void redo_move_effects(void)
   for (entry = &move_effect_journal[parent_top]; entry!=top_entry; ++entry)
   {
 #if defined(DOTRACE)
-    TraceValue("%u",entry-&move_effect_journal[0]);
+    TraceValue("%u",(unsigned int)(entry-&move_effect_journal[0]));
     TraceValue("%u",entry->type);
     TraceValue("%lu",entry->id);
     TraceEOL();
